Вакансии и переходы в них

+++ vacancy/models.py
@@ -6,7 +6,7 @@ from django.conf import settings
 class Vacancy(models.Model):
     title = models.CharField(max_length=255)
     description = models.TextField()
-    salary = models.DecimalField(max_digits=10, decimal_places=2)
+    salary = models.IntegerField()
     work_time = models.CharField(max_length=100)
     city = models.CharField(max_length=100)
     company = models.ForeignKey(Company, on_delete=models.CASCADE, related_name='vacancies')
